// src/app/model/userData.ts
export class userData {
  firstName: string = "";
  lastName: string = "";
  email: string = "";
  contactNo: number = null;
  age: number = null;
  state: string = "";
  country: string = "";
  address: string = "";
  address1: string = "";
  address2: string = "";
  companyAddress1: string = "";
  companyAddress2: string = "";
  tags: string = "";
  newsLetter: boolean;
  profilePhoto: string = "";
}